refactor: update environment variable imports to new structure and add validation schemas
