+++ core/src/db/db.prod.ts
@@ -4,7 +4,7 @@ import { neon } from "@neondatabase/serverless";
 import { drizzle, type NeonHttpDatabase } from "drizzle-orm/neon-http";
 
 import * as schema from "@/db/schema";
-import { SERVER_ENV } from "@/lib/env.server";
+import { SERVER_ENV } from "@/zap/env/env.server";
 
 const client = neon(SERVER_ENV.DATABASE_URL);
 export const db: NeonHttpDatabase<typeof schema> = drizzle({ client, schema });
