+++ core/src/zap-old/lib/polar/server.ts
@@ -2,7 +2,7 @@ import "server-only";
 
 import { Polar } from "@polar-sh/sdk";
 
-import { SERVER_ENV } from "@/lib/env.server";
+import { SERVER_ENV } from "@/zap/env/env.server";
 
 export const polarClient = new Polar({
   accessToken: SERVER_ENV.POLAR_ACCESS_TOKEN,
