+++ core/src/zap/errors/utils.ts
@@ -3,9 +3,9 @@ import "server-only";
 import type { z } from "zod";
 import { ZodError } from "zod";
 
-import { DEV } from "@/lib/env.public";
 import { BaseError, InternalServerError } from "@/zap/api/errors";
 import { logError } from "@/zap/api/logger";
+import { DEV } from "@/zap/env/env.public";
 import { generateUuid } from "@/zap-old/lib/crypto/utils";
 
 export type HandlerFunction<T extends unknown[], R> = (
